StoryNumber as decimal, OrderBy requires conversion, update references

+++ ComicsLibrary/EditModels/ReportEditModel.cs
@@ -10,7 +10,7 @@ namespace ComicsLibrary.EditModels
     {
         public string Title { get; set; }
         public string OriginalTitle { get; set; }
-        public int? StoryNumber { get; set; }
+        public decimal? StoryNumber { get; set; }
         public string ExtraInfo { get; set; }
         public string StoryType { get; set; }
         public string BookType { get; set; }
