+++ ComicsStore.MiddleWare/Models/Output/StoryOnlyOutputModel.cs
@@ -3,7 +3,7 @@
     public class StoryOnlyOutputModel : BasicOutputModel
     {
         public string StoryType { get; set; }
-        public int? StoryNumber { get; set; }
+        public decimal? StoryNumber { get; set; }
         public double? Pages { get; set; }
         public string ExtraInfo { get; set; }
         public string Language { get; set; }
